+++ profile-bee/src/lib.rs
@@ -35,5 +35,8 @@ pub mod probe_spec;
 
 pub mod v8;
 
+pub mod java;
+pub mod perf_map;
+
 mod trace_handler;
 pub use trace_handler::*;